# Process dependent files