--- a/e2e/process-dependent-files/expected-output-3.diff
+++ b/e2e/process-dependent-files/expected-output-3.diff
@@ -1,0 +1,1 @@
+[OK] Rector is done!