--- a/e2e/process-dependent-files/rector.php
+++ b/e2e/process-dependent-files/rector.php
@@ -1,0 +1,12 @@
+<?php
+
+use Rector\Config\RectorConfig;
+use Rector\TypeDeclaration\Rector\ClassMethod\NumericReturnTypeFromStrictReturnsRector;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->paths([
+        __DIR__ . '/src',
+    ]);
+
+    $rectorConfig->rule(NumericReturnTypeFromStrictReturnsRector::class);
+};