--- a/e2e/process-dependent-files/src/DependencyTestClass.php
+++ b/e2e/process-dependent-files/src/DependencyTestClass.php
@@ -1,0 +1,9 @@
+<?php
+
+class DependencyTestClass
+{
+    public function calledFunction(int $a)
+    {
+        return $a;
+    }
+}