--- a/e2e/process-dependent-files/src/DependencyTestClass.php.orig
+++ b/e2e/process-dependent-files/src/DependencyTestClass.php.orig
@@ -1,0 +1,10 @@
+<?php
+//DependencyTestClass.php
+
+class DependencyTestClass
+{
+    public function calledFunction(int $a): int
+    {
+        return $a;
+    }
+}