--- a/e2e/process-dependent-files/src/DependentTestClass.php
+++ b/e2e/process-dependent-files/src/DependentTestClass.php
@@ -1,0 +1,10 @@
+<?php
+
+class DependentTestClass
+{
+    public function callFunction()
+    {
+        $dependencyTestClass = new DependencyTestClass();
+        return $dependencyTestClass->calledFunction(1);
+    }
+}