--- a/e2e/process-dependent-files/src/UnrelatedTestClass.php
+++ b/e2e/process-dependent-files/src/UnrelatedTestClass.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+class UnrelatedTestClass
+{
+}