--- a/phpstan.neon
+++ b/phpstan.neon
@@ -114,6 +114,7 @@ parameters:
                 - rules/DeadCode/NodeManipulator/LivingCodeManipulator.php
                 - rules/EarlyReturn/Rector/If_/ChangeAndIfToEarlyReturnRector.php
                 - src/PhpParser/NodeTransformer.php
+                - src/FileSystem/FilesFinder.php
 
         # mapper re-use
         - '#Parameter \#1 \$type of method Rector\\PHPStanStaticTypeMapper\\TypeMapper\\ObjectWithoutClassTypeMapper\:\:mapToPhpParserNode\(\) expects PHPStan\\Type\\ObjectWithoutClassType, PHPStan\\Type\\Accessory\\Has(Property|Method)Type given#'