--- a/src/Application/ApplicationFileProcessor.php
+++ b/src/Application/ApplicationFileProcessor.php
@@ -8,6 +8,7 @@ use Nette\Utils\FileSystem as UtilsFileSystem;
 use PHPStan\Parser\ParserErrorsException;
 use Rector\Application\Provider\CurrentFileProvider;
 use Rector\Caching\Detector\ChangedFilesDetector;
+use Rector\Caching\FileDependenciesCache;
 use Rector\Configuration\Option;
 use Rector\Configuration\Parameter\SimpleParameterProvider;
 use Rector\FileSystem\FilesFinder;
@@ -52,12 +53,17 @@ final class ApplicationFileProcessor
         private readonly FileProcessor $fileProcessor,
         private readonly ArrayParametersMerger $arrayParametersMerger,
         private readonly MissConfigurationReporter $missConfigurationReporter,
+        private readonly FileDependenciesCache $fileDependenciesCache,
     ) {
     }
 
     public function run(Configuration $configuration, InputInterface $input): ProcessResult
     {
-        $filePaths = $this->filesFinder->findFilesInPaths($configuration->getPaths(), $configuration);
+        $allFiles = $this->filesFinder->findFilesInPaths($configuration->getPaths(), $configuration);
+        $filePaths = $this->filesFinder->filterUnchangedFiles($allFiles);
+
+        $filePaths = [...$filePaths, ...$this->findDependentFiles($allFiles, $filePaths)];
+
         $this->missConfigurationReporter->reportVendorInPaths($filePaths);
         $this->missConfigurationReporter->reportStartWithShortOpenTag();
 
@@ -94,9 +100,16 @@ final class ApplicationFileProcessor
         }
 
         if ($configuration->isParallel()) {
+            $this->fileDependenciesCache->cacheAllFiles($allFiles);
             $processResult = $this->runParallel($filePaths, $input, $postFileCallback);
         } else {
-            $processResult = $this->processFiles($filePaths, $configuration, $preFileCallback, $postFileCallback);
+            $processResult = $this->processFiles(
+                $filePaths,
+                $configuration,
+                $allFiles,
+                $preFileCallback,
+                $postFileCallback
+            );
         }
 
         $processResult->addSystemErrors($this->systemErrors);
@@ -108,20 +121,28 @@ final class ApplicationFileProcessor
 
     /**
      * @param string[] $filePaths
+     * @param string[] $allFiles
      * @param callable(string $file): void|null $preFileCallback
      * @param callable(int $fileCount): void|null $postFileCallback
      */
     public function processFiles(
         array $filePaths,
         Configuration $configuration,
+        ?array $allFiles = null,
         ?callable $preFileCallback = null,
-        ?callable $postFileCallback = null
+        ?callable $postFileCallback = null,
     ): ProcessResult {
         /** @var SystemError[] $systemErrors */
         $systemErrors = [];
 
         /** @var FileDiff[] $fileDiffs */
         $fileDiffs = [];
+
+        if ($allFiles === null) {
+            $allFiles = $this->fileDependenciesCache->getAllFiles();
+        }
+
+        $allFiles = array_fill_keys($allFiles, true);
 
         foreach ($filePaths as $filePath) {
             if ($preFileCallback !== null) {
@@ -131,7 +152,7 @@ final class ApplicationFileProcessor
             $file = new File($filePath, UtilsFileSystem::read($filePath));
 
             try {
-                $fileProcessResult = $this->processFile($file, $configuration);
+                $fileProcessResult = $this->processFile($file, $allFiles, $configuration);
 
                 $systemErrors = $this->arrayParametersMerger->merge(
                     $systemErrors,
@@ -161,11 +182,43 @@ final class ApplicationFileProcessor
         return new ProcessResult($systemErrors, $fileDiffs);
     }
 
-    private function processFile(File $file, Configuration $configuration): FileProcessResult
+    /**
+     * @param string[] $possibleDependentFiles
+     * @param string[] $filePaths
+     * @return string[]
+     */
+    private function findDependentFiles(array $possibleDependentFiles, array $filePaths): array
+    {
+        $filesToAnalise = array_flip($filePaths);
+
+        $dependentFiles = [];
+        foreach ($possibleDependentFiles as $possibleDependentFile) {
+            if (! isset($filesToAnalise[$possibleDependentFile])) {
+                $fileDependencies = $this->fileDependenciesCache->getFileDependencies($possibleDependentFile);
+
+                if ($fileDependencies !== null) {
+                    foreach ($fileDependencies as $fileDependency) {
+                        if (isset($filesToAnalise[$fileDependency])) {
+                            $dependentFiles[] = $possibleDependentFile;
+                            $this->changedFilesDetector->invalidateFile($possibleDependentFile);
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+
+        return $dependentFiles;
+    }
+
+    /**
+     * @param array<string,true> $allFiles
+     */
+    private function processFile(File $file, array $allFiles, Configuration $configuration): FileProcessResult
     {
         $this->currentFileProvider->setFile($file);
 
-        $fileProcessResult = $this->fileProcessor->processFile($file, $configuration);
+        $fileProcessResult = $this->fileProcessor->processFile($file, $allFiles, $configuration);
 
         if ($fileProcessResult->getSystemErrors() !== []) {
             $this->changedFilesDetector->invalidateFile($file->getFilePath());