--- a/src/Caching/Enum/CacheKey.php
+++ b/src/Caching/Enum/CacheKey.php
@@ -18,4 +18,14 @@ final class CacheKey
      * @var string
      */
     public const FILE_HASH_KEY = 'file_hash';
+
+    /**
+     * @var string
+     */
+    public const FILE_DEPENDENCIES_KEY = 'file_dependencies';
+
+    /**
+     * @var string
+     */
+    public const ALL_FILES_KEY = 'all_files';
 }