--- a/src/Console/Command/WorkerCommand.php
+++ b/src/Console/Command/WorkerCommand.php
@@ -139,6 +139,7 @@ final class WorkerCommand extends Command
             $processResult = $this->applicationFileProcessor->processFiles(
                 $filePaths,
                 $configuration,
+                null,
                 $preFileCallback
             );