--- a/src/DependencyInjection/LazyContainerFactory.php
+++ b/src/DependencyInjection/LazyContainerFactory.php
@@ -10,6 +10,7 @@ use Illuminate\Container\Container;
 use PhpParser\Lexer;
 use PHPStan\Analyser\NodeScopeResolver;
 use PHPStan\Analyser\ScopeFactory;
+use PHPStan\Dependency\DependencyResolver;
 use PHPStan\File\FileHelper;
 use PHPStan\Parser\Parser;
 use PHPStan\PhpDoc\TypeNodeResolver;
@@ -326,6 +327,7 @@ final class LazyContainerFactory
         NodeScopeResolver::class,
         ReflectionProvider::class,
         CachingVisitor::class,
+        DependencyResolver::class,
     ];
 
     /**