--- a/src/Util/FileHasher.php
+++ b/src/Util/FileHasher.php
@@ -40,6 +40,18 @@ final class FileHasher
         return $configHash;
     }
 
+    public function resolvePath(string $filePath): string
+    {
+        /** @var string|false $realPath */
+        $realPath = realpath($filePath);
+
+        if ($realPath === false) {
+            return $filePath;
+        }
+
+        return $realPath;
+    }
+
     private function getAlgo(): string
     {
         //see https://php.watch/articles/php-hash-benchmark