When using the cache, process dependent files of modified files to make sure we have covered all possible needed changes. See https://github.com/rectorphp/rector/issues/8801

This is implemented by using PHPStan's `DependencyResolver`. When we analyse a file, we save all the dependencies of this file in the cache. When we do a later analysis using the cache, we use these lists of dependencies to find the dependent files for any analysed file.

The implementation is quite simple and quite performant, I have not been able to measure a noticeable degradation in performance. Seems to cover all the cases that I could think of, probably needs more eyes on it and more testing to make sure everything works as supposed.

Includes a new end 2 end test for this functionality

See the PR comments for more information