@TomasVotruba 2024-08-27
Thanks for the PR. This is rather complex to review. Is there a way to use more from PHPStan to delegate the low-lever work?

Could you share few performance comparison before/after?


@staabm This is rather outside my area of focus, so I'd like to ask you for advisory opinion on this one :pray: