@staabm 2024-08-27
the caching on this level is a super-complex topic, therefore I would suggest to make it a opt-in feature before it stabilized.

---

let me write down a few corner stones from the top of my head, which are fundamental for the PHPStan caching.

- caching is disabled when `--debug` is used on the cli
- caching is disabled when passing the files to process via cli args, instead of using the actual config analyze-path
- phpstan outputs info about whether cache is used or not when `-vv` is used
- validates whether the cache is corrupt
- cache not used when older then x days
- cache needs to be invalidated on config changes (maybe with a few exceptions)
- AFAIR cache is only written from the master/main process. workers only read it.
- the `FileDependenciesCache` likely needs a cache-key which can be used to invalidate caches on fundamental rector-internal changes which turn a future rector release incompatible with caches already existant from previous rector versions 

Implementation details can be found in PHPStan `ResultCacheManager`