@carlos-granados 2024-08-27
Regarding performance, this PR does not aim to increase performance, in fact it will decrease it a little since it needs to calculate the dependencies for each file that it processes and then we need to use this data to calculate the dependent files for the files that need to be processed. I tried to measure the loss in performance when doing a full analysis with an empty cache and was not able to measure any real difference when using parallel processing, the uncertainty in the parallel processing seems to be bigger than the loss of performance that we are trying to measure. Not using parallel I got these results:
Without this code:
`bin/rector --clear-cache  96.95s user 3.27s system 97% cpu 1:42.73 total`
With this code:
`bin/rector --clear-cache  102.15s user 4.39s system 97% cpu 1:49.03 total`
So, there is a performance loss of around 5%
Running rector with a full cache and no files being processed shows these results:
Without this code:
`bin/rector  0.52s user 0.38s system 60% cpu 1.472 total`
With this code:
`bin/rector  0.54s user 0.45s system 54% cpu 1.808 total`
So, again there is a small performance loss but it will not be really noticeable
All these tests were done running Rector against the rector-src code base