@carlos-granados 2024-08-27
Regarding the Rector cache, right now it is a very simple cache, much simpler than the one used by PHPStan, it just saves the hash of any file processed so that this can be used later to find out if a file has changed. PHPStan saves a lot more data, including the dependencies for each file.
I could have modified the existing Rector cache so that for each file we saved both the hash and the dependencies but the end code would have been more complex than what I did, where the dependencies are cached separately. By caching them separately we don't have to touch the existing cache which will continue working in the same way.
I did not touch the code that clears the cache so it will continue to be cleared in the same cases that it was cleared before.
The `FileDependenciesCache` already has something that could be considered a `cache-key` as it uses the string `'file_dependencies' as a suffix for the hash used to cache the data, but I guess I could change it to `'file_dependencies_v1' to make it more explicit that this can be used to invalidate caches in future versions
Regarding making this an opt-in feature, I agree this could be good until this has stabilised. @TomasVotruba what about adding something like `->withProcessDependantFiles()` to enable this feature?