@carlos-granados 2024-08-27
Regarding using more of PHPStan, we cannot use its full caching capabilities because their cache is much more complex than the one used by Rector. The only possible piece of code that we could have maybe reused is the one that finds the dependent files from the saved dependencies but PHPStan has not extracted this code in a class or function that can be reused, it is embedded in the result cache manager code