@TomasVotruba 2024-08-27
I'll be prompt to avoid investing more time on your side. I share similar view as @staabm .
At the moment we have pretty good caching and it's good balance between complexity and speed. While this feature might catch 1-2 more files, it would make the future maintenance very costly and slow. 

Also, depedent files do not solve the need for re-run Rector in case of adding types. E.g. when new `string` type is added in single method the other dependent file still see it as file without a type as PHPStan often hold old types it parsed initially. Best practise is to re-run Rector in such cases, to catch them all in one PR.

I do the same for PHPStan and coding standard tools as well as they have the very same design.

Saying that, I want to keep current state of well balance.

Thank you @carlos-granados  :pray: