@carlos-granados 2024-08-27
@TomasVotruba no problem, just wanted to point out something. You said `Best practise is to re-run Rector in such cases, to catch them all in one PR.` The problem is that with the current functionality re-running Rector does not solve the problem unless you don't use the cache. So it happens that the results provided by Rector when you use the cache and when you don't use it can be different and I think that is a bad thing and what this PR aimed to solve