@carlos-granados 2024-10-04
If anyone else is interested in this option, it is available in this fork: https://github.com/carlos-granados/rector